// convex/_generated/api.d.ts
export declare const api: {
  games: {
    getGameState: string;
    getGameEvents: string;
    getPublicGames: string;
    getCommunityGames: string;
    getGamePresence: string;
    getTimerState: string;
    updateGameState: string;
    addGameEvent: string;
    controlTimer: string;
    joinGame: string;
    leaveGame: string;
    deleteGameEvent: string;
  };
};

export declare const internal: {};